requirements.txt
pyinstaller==5.12.0
pywin32==306
pillow